% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wqp.R
\name{load_wqp}
\alias{load_wqp}
\title{Load data directly from EPA WQP datawarehouse portal by URL.}
\usage{
load_wqp(start_date, end_date)
}
\arguments{
\item{start_date}{A start date for filtering results.}

\item{end_date}{An end date for filtering results.}
}
\value{
A dataframe from WQP with custom named columns.
}
\description{
Load data directly from EPA WQP datawarehouse portal by URL.
}
\examples{
my_data <- loadwqp("12-31-2015","12-31-2020")
}
