% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previews.R
\name{preview_locations}
\alias{preview_locations}
\title{Plot locations of monitoring values with ggmap()}
\usage{
preview_locations(wqp)
}
\arguments{
\item{wqp}{A dataframe from the WQP datawarehouse.}
}
\value{
A ggmap object of Hillsborough County, Florida with monitoring locations.
}
\description{
Plot locations of monitoring values with ggmap()
}
\examples{
preview_locations(wqp)
}
