% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previews.R
\name{preview_na}
\alias{preview_na}
\title{List number and percentages of NA for all variables.}
\usage{
preview_na(wqp)
}
\arguments{
\item{wqp}{A dataframe from the WQP datawarehouse.}
}
\value{
A dataframe of NA search results.
}
\description{
List number and percentages of NA for all variables.
}
\examples{
preview_na(wqp)
}
